backtrack somewhat (avoid size_t) in mkread and mkwrite

+++ inc/miriad.h
@@ -256,8 +256,8 @@ void xysetpl_c (int tno, int naxis, Const int *axes);
 /* maskio.c */
 char *mkopen_c (int tno, char *name, char *status);
 void mkclose_c (char *handle);
-int  mkread_c  (char *handle, int mode, int *flags, off_t offset, size_t n, size_t nsize);
-void mkwrite_c (char *handle, int mode, Const int *flags, off_t offset, size_t n, size_t nsize);
+int  mkread_c  (char *handle, int mode, int *flags, off_t offset, int n, int nsize);
+void mkwrite_c (char *handle, int mode, Const int *flags, off_t offset, int n, int nsize);
 void mkflush_c (char *handle);
